refactor: Relocate rental detail components to presentation layer (#203)
