+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/components/section/RentalPaymentSection.kt
@@ -8,7 +8,7 @@ import com.example.rentit.R
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.presentation.rentaldetail.model.PriceSummaryUiModel
 import com.example.rentit.common.component.TitledContainer
-import com.example.rentit.common.component.PriceSummary
+import com.example.rentit.presentation.rentaldetail.components.PriceSummary
 
 @Composable
 fun RentalPaymentSection(
